refactor: split resource executor backends
