+++ src/Runtime/devenv/pkg/resource/executor/observer_test.go
@@ -1,4 +1,4 @@
-package resource
+package executor
 
 import "testing"
