+++ src/Runtime/devenv/pkg/resource/executor/outputs.go
@@ -1,9 +1,10 @@
-package resource
+package executor
 
 import (
 	"sync"
 
 	"altinn.studio/devenv/pkg/container/types"
+	"altinn.studio/devenv/pkg/resource"
 )
 
 // Output contains runtime facts produced while applying a resource.
@@ -22,20 +23,21 @@ type ContainerOutput struct {
 	HostPorts   []types.PublishedPort
 }
 
-type noOutput struct{}
+// NoOutput marks a successful apply operation that produces no runtime output.
+type NoOutput struct{}
 
 // Outputs provides typed access to runtime outputs keyed by resource ID.
 type Outputs struct {
-	values map[ResourceID]Output
+	values map[resource.ResourceID]Output
 }
 
 type outputStore struct {
-	m  map[ResourceID]Output
+	m  map[resource.ResourceID]Output
 	mu sync.RWMutex
 }
 
 func newOutputStore() *outputStore {
-	return &outputStore{m: make(map[ResourceID]Output)}
+	return &outputStore{m: make(map[resource.ResourceID]Output)}
 }
 
 func (o ImageOutput) clone() Output {
@@ -56,12 +58,12 @@ func (o ContainerOutput) withClone() ContainerOutput {
 	return cloned
 }
 
-func (noOutput) clone() Output {
-	return noOutput{}
+func (NoOutput) clone() Output {
+	return NoOutput{}
 }
 
 // Get returns the raw output for a resource ID.
-func (o Outputs) Get(id ResourceID) (Output, bool) {
+func (o Outputs) Get(id resource.ResourceID) (Output, bool) {
 	output, ok := o.values[id]
 	if !ok {
 		return nil, false
@@ -70,7 +72,7 @@ func (o Outputs) Get(id ResourceID) (Output, bool) {
 }
 
 // Image returns the typed image output for a resource ID.
-func (o Outputs) Image(id ResourceID) (ImageOutput, bool) {
+func (o Outputs) Image(id resource.ResourceID) (ImageOutput, bool) {
 	output, ok := o.values[id]
 	if !ok {
 		return ImageOutput{}, false
@@ -83,7 +85,7 @@ func (o Outputs) Image(id ResourceID) (ImageOutput, bool) {
 }
 
 // Container returns the typed container output for a resource ID.
-func (o Outputs) Container(id ResourceID) (ContainerOutput, bool) {
+func (o Outputs) Container(id resource.ResourceID) (ContainerOutput, bool) {
 	output, ok := o.values[id]
 	if !ok {
 		return ContainerOutput{}, false
@@ -102,17 +104,17 @@ func (o Outputs) Len() int {
 
 func (s *outputStore) Reset() {
 	s.mu.Lock()
-	s.m = make(map[ResourceID]Output)
+	s.m = make(map[resource.ResourceID]Output)
 	s.mu.Unlock()
 }
 
-func (s *outputStore) Set(id ResourceID, output Output) {
+func (s *outputStore) Set(id resource.ResourceID, output Output) {
 	s.mu.Lock()
 	s.m[id] = output.clone()
 	s.mu.Unlock()
 }
 
-func (s *outputStore) Get(id ResourceID) (Output, bool) {
+func (s *outputStore) Get(id resource.ResourceID) (Output, bool) {
 	s.mu.RLock()
 	output, ok := s.m[id]
 	s.mu.RUnlock()
@@ -122,7 +124,7 @@ func (s *outputStore) Get(id ResourceID) (Output, bool) {
 	return output.clone(), true
 }
 
-func (s *outputStore) Image(id ResourceID) (ImageOutput, bool) {
+func (s *outputStore) Image(id resource.ResourceID) (ImageOutput, bool) {
 	output, ok := s.Get(id)
 	if !ok {
 		return ImageOutput{}, false
@@ -134,7 +136,7 @@ func (s *outputStore) Image(id ResourceID) (ImageOutput, bool) {
 	return image.withClone(), true
 }
 
-func (s *outputStore) Container(id ResourceID) (ContainerOutput, bool) {
+func (s *outputStore) Container(id resource.ResourceID) (ContainerOutput, bool) {
 	output, ok := s.Get(id)
 	if !ok {
 		return ContainerOutput{}, false
@@ -150,7 +152,7 @@ func (s *outputStore) Snapshot() Outputs {
 	s.mu.RLock()
 	defer s.mu.RUnlock()
 
-	values := make(map[ResourceID]Output, len(s.m))
+	values := make(map[resource.ResourceID]Output, len(s.m))
 	for id, output := range s.m {
 		values[id] = output.clone()
 	}
@@ -158,6 +160,16 @@ func (s *outputStore) Snapshot() Outputs {
 }
 
 func isNoOutput(output Output) bool {
-	_, ok := output.(noOutput)
+	_, ok := output.(NoOutput)
 	return ok
+}
+
+func clonePublishedPorts(ports []types.PublishedPort) []types.PublishedPort {
+	if len(ports) == 0 {
+		return nil
+	}
+
+	cloned := make([]types.PublishedPort, len(ports))
+	copy(cloned, ports)
+	return cloned
 }
