+++ src/Runtime/devenv/pkg/resource/executor/plan_test.go
@@ -1,4 +1,4 @@
-package resource
+package executor
 
 import (
 	"slices"
